third_party/blink/web_tests/virtual/scoped-custom-element-registry-disabled/external/wpt/custom-elements/scoped-registry/scoped-registry-define-upgrade-criteria.tentative-expected.txt

This is a testharness.js-based test.
[FAIL] Adding definition to global registry should not affect shadow roots using scoped registry
  Illegal constructor
[PASS] Adding definition to global registry should affect shadow roots also using global registry
[FAIL] Adding definition to scoped registry should affect all associated shadow roots
  Illegal constructor
[FAIL] Adding definition to scoped registry should not affect document tree scope
  Illegal constructor
[FAIL] Adding definition to scoped registry should not affect shadow roots using other registries
  Illegal constructor
[FAIL] Adding definition to global registry should not upgrade nodes no longer using the registry
  Illegal constructor
[FAIL] Adding definition to scoped registry should not upgrade nodes no longer using the registry
  Illegal constructor
[FAIL] Adding definition to scoped registry affects associated shadow roots in all iframes
  Illegal constructor
[FAIL] Adding definition to scoped registry affects associated shadow roots in other frame trees
  assert_true: expected true got false
[FAIL] Adding definition to scoped registry should not upgrade disconnected elements
  Illegal constructor
[FAIL] Adding definition to scoped registry should not upgrade nodes in constructed documents
  Illegal constructor
[FAIL] Adding definition to scoped registry should not upgrade nodes in detached frames
  Illegal constructor
[PASS] Adding definition to scoped registry should not upgrade nodes in closed windows
Harness: the test ran to completion.
